codegen/schema: add bigint and sequence migrations
This change modifies codegen in the following important ways

All integer types now drop down to the underlying SQLAlchemy type to
correctly specify the field (as BigInteger e.g.) which translates
into the correct underlying database type.

Sequences are now supported on ints using the auto_update property.
this allows composite PKs with only part of the key being auto increment
or arbitrary fields if desired.

A migration has been added - it is mostly automatic but sequences
must be created and dropped in the migration file by hand. This is a
limitation of the underyling auto migration revision generation code. An
example is included.

The jobs test has been updated with additional tests for multiple
results
containing different data to ensure that result sequences are actually
generated correctly..