[api/app] First pass of jobs API and primary key codegen fixes (#182)
* First pass of jobs api

* Fix line ending differences causing large diff

* Resolve unit test issues

* codegen/schema: add bigint and sequence migrations

This change modifies codegen in the following important ways

All integer types now drop down to the underlying SQLAlchemy type to
correctly specify the field (as BigInteger e.g.) which translates
into the correct underlying database type.

Sequences are now supported on ints using the auto_update property.
this allows composite PKs with only part of the key being auto increment
or arbitrary fields if desired.

A migration has been added - it is mostly automatic but sequences
must be created and dropped in the migration file by hand. This is a
limitation of the underyling auto migration revision generation code. An
example is included.

The jobs test has been updated with additional tests for multiple
results
containing different data to ensure that result sequences are actually
generated correctly..

* fix FK definition for new sqa int types

* update SQLite codegen variant

SQLite does not support autoincrement on bigint values, they must be
treated as integers

* update the sqlite autoincrement behaviors

* fix codgen for unique constraints (composite pk) in sqlite

* add local helper script to run CI

* add unique constrait migration for job result seq

* uniqueconstraint in schema from codegen

* add overridable tz info to config

* santize and rationalize db timezone handling

* updating poetry.lock

* update pylint hints for table_args

* additional pylint cleanup

---------

Co-authored-by: kevin <kevin@mythica.ai>
Co-authored-by: Jacob Repp <jacobrepp@gmail.com>