[houdini-ui] Allow direct editing of int and float parms (#946)
The numbers can now be edited directly by simply clicking on them. Enabled for both ints and floats.
For floats, you can express numbers as 1.0e-1. It has validation to ensure only valid numbers are entered.
It will also respect min_is_strict and max_is_string on the parmTemplates. The slider stops at the min and max but the values can be manually adjusted higher or lower if and only if those flags are set.